+++ src/main/java/org/nrg/containers/services/impl/ContainerFinalizeServiceImpl.java
@@ -7,6 +7,7 @@ import com.google.common.collect.Maps;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.nrg.action.ClientException;
+import org.nrg.action.ServerException;
 import org.nrg.containers.api.ContainerControlApi;
 import org.nrg.containers.exceptions.ContainerException;
 import org.nrg.containers.exceptions.DockerServerException;
@@ -358,6 +359,13 @@ public class ContainerFinalizeServiceImpl implements ContainerFinalizeService {
                     final String message = prefix + "Could not upload files to resource.";
                     log.error(message);
                     throw new ContainerException(message, e);
+                }
+
+                try {
+                    catalogService.refreshResourceCatalog(userI, createdUri);
+                } catch (ServerException | ClientException e) {
+                    final String message = String.format(prefix + "Could not refresh catalog for resource %s.", createdUri);
+                    log.error(message);
                 }
             } else if (type.equals(ASSESSOR.getName())) {